FEST 63 : - add base Date assertions with String parameter : isIn, isNotIn, isEqualTo, isNotEqualTo - add isAfterYear assertion - complete unit tests
Genery isIn and isNotIn methodis with varrag parameter in Assert interface and AbstractAssert
It was not possible to generify isIn/isNotIn with Collection param because in some case it would result in calling isIn(A... values) instead of isIn(Collection<A> values).
For example, let's say frodo, sam and pipin are instances of Character class and fellowshipOfTheRing a List<Character> containing frodo.
It turns out that 'assertThat(frodo).isIn(fellowshipOfTheRing);' would result in calling isIn(Object... values) with one parameter of type List<Character> instead of calling isIn(Collection<Object>)
The reason is that List<Character> is not inheriting Collection<Object> because java collection are not covariant.