Fixes #238 : Can use extracting with SoftAssertions
Extracting was failing because the resulting Assert class was not
a proxy collecting errors.

This commit register a new method interceptor for cglib proxies: ProxifyExtractingResult.
If the intercepted method name contains "extracting", a filter choose to use ProxifyExtractingResult.
In all the other cases the interceptor used is the former ErrorCollector.