Ignore static methods when finding property accessors. Fixes #1458
Previously, when looking for property accessors, Introspection would consider
a static method with a name in the required form as being an accessor for a
property. Where such a static method existed, this was likely to result in the
wrong value being considered as the static method is unlikely to return the
correct value for an instance's property.

This commit updates Introspection so that static methods are ignored when
looking for property accessors.