Deprecate Java 6/Android assertions entry point. Fixes #1474.
The goal of these entry points was to provide Java6/Android compatible assertions but since we have moved to Java 8 it became more complicated, filtering out java 8 types is not sufficient as we have refactored some assertions implementation to use lambdas and streams that made them improper to use in Java 6. Reverting all this work is not worth it as Android is moving to have more java 8 support.