Introduce RegisterExtension for JUnit Jupiter
Prior to this commit, AssertJ offered JUnitJupiterSoftAssertions and
JUnitJupiterBDDSoftAssertions for working with SoftAssertions or
BDDSoftAssertions in JUnit Jupiter based tests. Those two extensions
unfortunately have several drawbacks as outlined below.

- lacking support for parallel test execution
- lacking support for @TestInstance(PER_CLASS) lifecycle semantics

In addition, JUnitJupiterSoftAssertions or
JUnitJupiterBDDSoftAssertions can easily be declared in a manner that
leads to undesirable behavior. For example, programmatic registration
via @RegisterExtension allows an instance of JUnitJupiterSoftAssertions
or JUnitJupiterBDDSoftAssertions to be used in constructors and
lifecycle methods, even though `assertAll()` would not be called for
the given JUnitJupiterSoftAssertions or JUnitJupiterBDDSoftAssertions
instance until after the test method has been executed. Similarly, with
@Nested test classes, programmatic registration via @RegisterExtension
allows an instance of JUnitJupiterSoftAssertions or
JUnitJupiterBDDSoftAssertions to be registered via a field in an
enclosing class, but use of that field is not technically supported in
a @Nested test class.

This commit addresses these shortcomings by introducing a new
SoftAssertionsExtension for use with JUnit Jupiter. This extension
supersedes the existing JUnitJupiterSoftAssertions and
JUnitJupiterBDDSoftAssertions extensions.

Specifically, SoftAssertionsExtension:

- supports the injection of an instance of SoftAssertions or
  BDDSoftAssertions into "testable" methods such as those annotated
  with @Test, @TestFacttory, @RepeatedTest, @ParameterizedTest, and
  @TestTemplate
- throws an exception if SoftAssertionsExtension is used with a test
  constructor or lifecycle method such as @BeforeEach and @AfterEach
- invokes `assertAll()` on the instance of SoftAssertions or
  BDDSoftAssertions immediately after the "testable" method has been
  executed (i.e., by implementing the AfterTestExecutionCallback API)

The SoftAssertionsExtension is tested using the JUnit Platform Test Kit.

Closes #1418