Add "it" infrastructure for integration tests
This commit adds an optional maven-invoker-plugin based profile
to the build configuration that builds a set of Maven projects
located in the newly added "src/it" directory.

No dedicated verification script is evaluated after such an
"it" project build for the time being. A successful build is
considered as successful test case.