Rework `StandardComparisonStrategy#areEqual` to avoid shortcuts (#2114)
The previous behavior relied on `java.util.Objects#deepEquals`, which
performs a reference check before invoking `equals` on the first
parameter. While this is not an issue for proper implementations of
the `equals` contract, it was behaving incorrectly in case of
implementations not satisfying the reflexive property. Also, performing
the reference check outside of the `equals` implementation reduced the
overall coverage.

The tests now provide thorough regression to ensure that no shortcut is
taken, especially in case of implementations violating the contract.