Support String assertions for InputStream
In some of my tests I have to check InputStreams which contain text. I
cannot check the whole test with assertThat(...).hasContent(...) because
the text contains non-predictable parts (e.g. UUIDs).

This commit adds a new method asString(Charset) for InputStream
assertions. It returns an AbstractStringAssert so that all assertions
for String can be used for the InputStream. Here is an example

    assertThat(new ByteArrayInputStream("abc".getBytes()))
      .asString(UTF_8)
      .startWith("a")
      .endsWith("c");

The method asString(Charset) is similar to the one of ByteArrayAssert.
The test is for the most part a copy of
ByteArrayAssert_asString_with_charset_Test