Fall back to `Map` copy when cloning causes any `RuntimeException`
This affects `contains` assertions searching for non-expected elements
(i.e., keys or entries) and allows proper executions with Hibernate's
`PersistentMap` and `PersistentSortedMap` types.

The tests are written against Hibernate 6.0.0.Beta3 as the way to
instantiate these types has been radically simplified since the last
stable version, but the general behavior is expected to be the same.

Fixes #2448.