[ISSUE 2629] Improve readability of containsExactly
Removes elementsDifferAtIndex() from assertContainsExactly() and re-use
shouldContainExactlyWithDiffAssertionError() in order to show the actual
and expected sets entirely. Making the differencesFound() check unnecessary.

It also introduces a new shouldContainExactly() and checks for both
the missing and unexpected diffs to be empty. Otherwise cases where the
actual and expected only differ in order, result into logging of missing
elements - which there are none.