GH-2647: Implement skeleton for stack trace assertions.
This introduces the skeleton for two new types of
assertion/assumption:

- assertThat(StackTraceElement) - to perform assertions
  upon StackTraceElement types.
- assertThat(StackTraceElement[]) - to perform assertions
  upon StackTraceElement[] types.

Due to the new signature conflicting with one of the existing
test suites, I have implemented a temporary workaround to prevent
having to totally reimplement that class, since the new
API is not yet usable in this set of changes. I will remove
this in the second PR that implements the actual functionality.

I have also made one of the proxy methods in the API less restrictive
to allow it to work with passing array types around.