build: Use Bnd feature to write resolved bndrun files to output folder
Bnd 6.4.0 allows writing and reading bndrun files from different
folders. So we change to use this feature so that the resolver plugin
writes the resolved bndrun file to the target folder and the testing
plugin uses the resolved bndrun file from the target folder.

This change means that running the maven build does not modify checked
in source (bndrun files).

Signed-off-by: BJ Hargrave <bj@hargrave.dev>