Add exception for usages of `isEqualTo` and `isNotEqualTo` on assertions
Before this change, the following code:

    AbstractAssert<?, ?> assertion = assertThat(something);
    assertThat(assertion).isEqualTo(assertion);

would throw an exception with a message about the use of `equals` being
unsupported, suggesting to use `isEqualTo` instead. The message is
somewhat confusing to the reader, since `isEqualTo` is indeed used, and
is due to the fact that `isEqualTo` internally relies on the `equals` of
`actual`, which is an `AbstractAssert` instance and throws the
aforementioned exception.

`isEqualTo` and `isNotEqualTo` now check if `actual` is an assertion
instance and raise an `UnsupportedOperationException` if so, with
a message suggesting to use `isSameAs` and `isNotSameAs` instead.

This is, for example, useful for testing custom assertion types for
extension libraries, where the use of `satisfies` instead of `isEqualTo`
may be desirable.

Closes #2921