Fixes a bug when failing assertions on DirectoryStream types (#3036)
The DirectoryStream class is somewhat special because the API allows it
to only support .iterator() being called at most once on any
implementations. Due to the nature of how the StandardRepresentation
currently works for iterable types, any assertion error that
is thrown as a result of assertions failing on a DirectoryStream
object can result in IllegalStateExceptions being propagated to
the caller if .iterator() has already been called on the class.
This can break further if using soft assertions, as the state of
the class could change unexpectedly resulting in different
exceptions being raised depending on the execution order.

I've added a mechanism to deal with allowing the definition of
so-called 'blacklisted' iterable types, just in case other
cases like this appear in the JRE standard library anywhere
in the future. This will prevent unexpected errors occuring
when dealing with the NIO FileSystem APIs.