Format large arrays as strings. Fix #3065
The process for converting an array of primitives to a string was:
1. convert the array to an array of objects
2. convert the objects to strings
3. select up to 1000 elements to print

This is fine if there aren't too many elements, but if there more than 2^29
elements, all the copying may result in an OOM exception.

With this change, instead of copying the arrays, create a collection
that proxies the original collection as a list of Strings. The strings
aren't created unless they're actually needed.

Testing:

big arrays were created with 2^28 elements as 2^30 would already raise an OOM on Joel's computer
We used 2^28 as we are not sure what are the memory settings of github actions