Format large arrays as strings
The process for converting an array of primitives to a string was:
1. convert the array to an array of objects
2. convert the objects to strings
3. select up to 1000 elements to print

This is fine if there aren't too many elements, but if there are 2^30
elements, all the copying may result in an OOM exception.

With this change, instead of copying the arrays, create a collection
that proxies the original collection as a list of Strings. The strings
aren't created unless they're actually needed.

Testing:

Originally, on my computer, trying to turn big arrays into strings
resulted in OOM error after about 10 seconds of computation. Now, this
completes in a fraction of a second.