Make test for long iterable smaller
Creating the test data for the long iterable causes an out of memory
error in the GitHub automated build system. This change uses a much
smaller example, with fewer things to iterate through and smaller items.
It adds a timeout, however, so the test still should fail without the
related change that makes this more efficient.