Treat empty differently in immutability checks
If a collection is empty, it doesn't matter what methods like clear()
and remove() do, when checking whether the collection is modifiable. As
long as it isn't possible to add anything, the collection will never
contain any elements.

This matters because throwing UnsupportedOperationException is optional.
If a collection is empty, remove() can either do nothing or throw
UnsupportedOperationException. So if the collection being tested for
unmodifiability is empty, calling remove() doesn't tell you anything
about whether the collection is modifiable.

The current approach is to have lists of pre-approved unmodifiable
collections, but this won't work for collections that aren't one of the
known collections but use no-op for things like remove().