DualValue avoid calling actual.hashCode() and expected.hashCode()
The hashCode of `DualValue` needs to match the object references of actual and expected, because that's what `equals` uses.

The current implementation caches the hashCode of both objects at construction, but this doesn't really match the requirement exactly.

The motivation for changing this is around avoiding calling hashCode unnecessarily on actual and expected objects, but I think it fixes potential bugs too. You could:
* Create a DualValue.
* Mutate a value in actual or expected that's present in the object's hashCode method.
* Create another DualValue.

Both of the DualValues would be equal to each other according to equals() but would have different hashcodes.

We can continue to cache the hashCode if you prefer, but better to generate it based on the object references using `System.identityHashCode`.