BREAKING CHANGE: Remove Function support when extracting from Iterable
In Kotlin the call to extracting with a lambda for an iterable assert is ambiguous. This is due to the extracting method existing for ThrowingExtractor and Function, which have the same effective signature, as checked exceptions do not exist in Kotlin. We opt to continue using the extracting method for ThrowingExtractor, since it is the more flexible functional interface, as it supports that the lambda provided can throw an exception.