Fix thread-safety in `AbstractDateAssert` (#3874)
97b642a0 introduced a regression that broke `AbstractDateAssert`
thread-safety by replacing synchronization on `DEFAULT_DATE_FORMATS`
with synchronization on the method that parses with the default
formats. Such a synchronization cannot work properly with multiple
instances of `AbstractDateAssert` because `DEFAULT_DATE_FORMATS` is
static.

This change fixes the regression by replacing the synchronization on
the method level with a `ThreadLocal` that guards access to the current
`DEFAULT_DATE_FORMATS` content.